- rework configuration and checks of the complex vmtest "allindata"
- added new regex based file content verification
- added testcase class variable "extra_kern_args" to be able to
  influence the kernel parameters of the installed system
- fix the handling of KEEP_VMTEST_DATA
- fix an issue with -curses on the backlevel qemu in trusty
  by passing '-vga std' to avoid getting cirrus card and framebuffer.
- enhancing README-vmtest.txt with my lessons learned
- provide an option to set the name parameter at mdadm create
- check for vnc port collisions and automatically assign alternatives

bzr-revno: 290

+++ tests/vmtests/__init__.py
@@ -134,13 +134,15 @@ class TempDir:
                               stdout=DEVNULL, stderr=subprocess.STDOUT)
 
     def __del__(self):
-        # remove tempdir
-        shutil.rmtree(self.tmpdir)
+        if (os.getenv('KEEP_VMTEST_DATA', "false") != "true"):
+            # remove tempdir
+            shutil.rmtree(self.tmpdir)
 
 
 class VMBaseClass:
     disk_to_check = {}
     fstab_expected = {}
+    extra_kern_args = None
 
     @classmethod
     def setUpClass(self):
@@ -163,6 +165,9 @@ class VMBaseClass:
             cmd.extend(["--silent", "--power=off"])
 
         cmd.extend(["--serial-log=" + self.install_log])
+
+        if self.extra_kern_args:
+            cmd.extend(["--append=" + self.extra_kern_args])
 
         # check for network configuration
         self.network_state = curtin_net.parse_net_config(self.conf_file)
@@ -319,10 +324,15 @@ class VMBaseClass:
         for f in files:
             self.assertTrue(os.path.exists(os.path.join(self.td.mnt, f)))
 
-    def check_file_content(self, filename, search):
+    def check_file_strippedline(self, filename, search):
         with open(os.path.join(self.td.mnt, filename), "r") as fp:
             data = list(i.strip() for i in fp.readlines())
         self.assertIn(search, data)
+
+    def check_file_regex(self, filename, regex):
+        with open(os.path.join(self.td.mnt, filename), "r") as fp:
+            data = fp.read()
+        self.assertRegex(data, regex)
 
     def get_blkid_data(self, blkid_file):
         with open(os.path.join(self.td.mnt, blkid_file)) as fp:
